Reword getAllPull comment and rename variables
It turns out that getAllPulls really does return a promise to an array
of promises, I overlooked the first promise due to missing the "then" in
app.js. x_x

Also, rename variables to be consistent with the naming convention in
getAllPulls. This is the convention I've read across some parts of the
internet regarding promises. I think promises should have a Promise
suffix because of the fact that they are not the object they claim to
be; they are a promise to that object which is interfaced with quite
differently than the object itself.