Insert signatures into DB one at a time
Now signatures will be inserted the same way whether we're starting
the app, or whether an `issue_comment` webhook was delivered. The
implications of this are that de'active'ating signatures is now done
when inserting a signature into the DB. (e.g. inserting an `un_dev_block`
signature will set all previous `dev_block` signatures in the DB to `active = 0`.
This allows us to pull all the signature logic out of the Pull object, which is
much cleaner, and we don't have to duplicate any logic when processing webhooks.