Move renderInto function to Templates and rename
Now called renderIntoContainer. Templates seems like a more logical
place for it, as one would only rarely want to render a template to a
template function and not to an inserted DOM node. Thus, this
functionality seems like it should be generally available, rather than
just available in the Column class.