Intialize webhook-compelte promise to a resolved promise
This solves an issue where we didn't handle every type of update type. github
added a new "labeled" type. We could just make a default case in the switch
statement, but we felt it was cleaner to just initialize the promise
before the switch is entered, and then whichever hooks need to override it
can do so.