Add Labels to database on restart
Delete all of a pull's labels from the database then insert the current ones.

Also fix a small a small misnomer with labels. In the database we call what
github calls the label `name` the label `title`. Use the same terminology that
we use in the database in the code.