Add link to cryo-storage github page
The "n frozen pulls" text in the navbar seems like it ought to
be a link. So let's make it one.

This is my first code contribution to pulldasher!! Woo!

fixes #122

Dynamically create the link

Instead of hardcoding it. This means it'll work if the repo changes.

Style

My spacing was off in a string thing

Use dot notation on the objects

Instead of the nested array notation.