Format Dates in Safari
We currently define formatting options in our call to
`Date.toLocaleDateString()`, which shortens the month name to three characters
and displays the day of the month (Sep 15). Unfortunately, Safari does not
support these formatting options, so the entire date is shown (September 15,
2014).

We now perform a regular expression on the date string to get the correct
formatting.