Add easy way to copy branch names from pulldasher
Fixes #148

I like the clipboard icon where it is. It'd be easy to make it only appear on hover, relocate, or do any number of things (probably need to get a designer in on it?).
Hover and it prints the branch name. Click it brings up a prompt with nice, selected text, all ready for you.