Make pull signatures get recomputed on update
When a pull is updated, we need to recompute the list of signatures that
are current, old, and user. This wasn't happening, due to updates being
handled through the update method rather than running the constructor
again. This led to pull signoff marks being inaccurate. Now the
constructor just calls update, since update does exactly what we wanted
to do in the constructor anyway.