Support multiple filter functions in config
We can now use an array of functions rather than just one function in the config
for filtering, sorting, and grouping. The functions will be chained, with the
first receiving the full list of data, then the second receiving the filtered
list resulting from the first, then the third receiving the filtered list
resulting from the second, and so on. The same thing will happen with sorting
and grouping functions, but the results are probably more or less meaningless:
unless we have a stable sort, for example (which I don't believe we do),
multiple sorting won't be guaranteed to mean anything.