Example config: update to handle unicode emojis as well
Github started making their emoji auto-completor inject unicode emojis
instead of ascii representations like :+1: or :smile:. This updates
the regex to detect unicode ranges that contain emojis. The regex is a
bit more permissive than needed, but the set symbols that are used
as emojis is dispersed across a smattering of unicode ranges.