Github: obey the rate limit!
All hail the rate limit!

This adds some smarts to our github usage and will allow requests to
go through unimpeded until we get to the choke point, at which point
we start throttling requests such that we stay *just* under the rate
limit.

Obviously, not knowing the future means we can't be all that smart
about it.

The delay is such that if you have an unlimited number of requests
coming in, we'll stay jsut below the rate limit. If you reach the
choke point but only have a few more requests, they will be throttled
unnecessarily, but we'll just have to live with that.

This means that now, the rebuild-history option will totally work
without hitting the rate limit.