Issues: made refresh module for refreshing logic
This module uses an async queue to execute one issue refresh at a
time, instead of all in parallel. This is important for when we
refresh all issues and end up with thousands of them.

It ensures also that the same issue can't be refreshed twice in
parallel (which would have undefined behavior).

Will likely do the same thing with pulls later on.

Also, add bin/refresh-issue script.