Github Comment events: deal with edit / delete
Now we can finally deal with these new event types.

Since pull request review comments can't have signatures,
we only need to delete or update the comment.

On the other hand, issue-level comments on pulls can hav signatures
and if we edit a comment to introduce or remove a signature, that
could have large affects on the rest of the signatures and there's no
easy way to undo the changes that an un_dev_block signature has done.
The easiest wya to deal with it is to completely refresh the given
pull. It's a litle overkill, but it's very little code and it's easy
to understand.