Debug Log: allow disabling of timestamp
When you're logging the output through any modern logging
system (syslog, journalctl, ...) a timestamp is prepended anyway.
There's no point in adding a second timestamp.

Let's make that configurable using an ENV variable. We wrap
the debug module and override the formatArgs function to
not include the timestamp.

Note: we dropped the colors in this mode cause we don't care.