Issue labels: allow config to specify conversion
We had a weird balance between flexibility (field name + regex)
and hard-coding of logic; it required custom processing for the one
case where we used it. Let's just make that deterministic function
part of the config.

Also, upon every processing, reset each field to null so that if
the label disappears, the value does too.