Refresh queue: always call next() when done.
We allow queuers to put a function on the refresh queue
to be notified when the queue has been executed to that
point. We had forgetten to still proceed to the next
queued item after that signla function had been called.

Now, we call next() for both the normal case and the
signal function case.

This effectively meant that only one issue would ever
really be refreshed before the queue stopped going.