Pull requests: deal with label renames
We wrote some code so that label events that happened across a rename:

   1. label added "LabelA"
   (renamed to "labela")
   2. label removes "labela"

wouldn't leave the issue in the DB with the LabelA just cause it was
removed after it had been renamed. We use ghIssue.labels as the
canonical list and filter the events by that. The problem is that pull
requests don't have a list of labels, so that code was being skipped.

The solution is to get the issue of the same number (every pull
request is an issue at heart) and use it for it;s list of labels.