Add Dockerfile
This adds a Dockerfile that allows running pulldasher in a docker
container. The container does not include an instance of mysqld, so the
pulldasher container will not be able to use a connection to mysql using
localhost. It will have to be able to connect to a mysqld server running
in either another container, or another ip address.