github api responses: don't depend on user.login
Sometimes github user accounts are deleted.

When they are included in an api response, the user looks like:

Normally:

    issue: {
        user : {
            login: "danielbeardsley",
            url: ...
        }
    }

When an response refers to a deleted user, the value is null:

    issue: {
        user : null
    }

This change wraps all places where we access .login from a user
object in an api response.

Most places would have previously crashed the app. The one that
changed now produces "deleted user" instead of NULL. This
allows the DB columns to still be NOT NULL. Keeping them from
being null makes it so we don't have to deal with nulls in the
rest of our code.

Perhaps in the future we'll want to do something different, but
this is safe for now.