Add a migration script and entrypoint.sh
This adds a new script called migrate.js, which ensures that there is an
existing table structure for pulldasher. This could probably be more
complex, but for now it will create the necessary tables if the database
does not have them, and it will do nothing if the tables already exist.

This lets us run the migrate script multiple times without destroying an
already configured database.