Update npm dependency versions
This bumps the versions of middleware and standalone npm packages that we use
throughout pulldasher.

This only covers low-hanging fruit modules that don't have breaking changes.
Once example of a module that isn't getting updated in this pull is Express,
because moving from v3 to v4 is going to require some code changes.