Move var declaration to top of function
This was caught by the updated version of JSHint. We are assigning the result
of a setTimeout() after using it in an event handler we pass to socket.io.

I don't think this is guaranteed to fail, because the variable will most likely
be assigned by the time the event handler is called, but it's better to make
sure that the variable is actually set, and not make a race condition between
reaching the end of the function and having an event handler be run.