Handle passport authentication errors correctly
We were passing both a success and error callback to `passport.authenticate`,
which doesn't match the way that that function handles errors.
`passport.authentiate` passes an error object as the first parameter to the
first callback which we now check before operating on the response from the
OAuth authentication.

Based on the passport source, the second callback was being clobbered when the
second argument was also a callback.