Redirect to original URL after OAuth success
Previously if you accessed pulldasher using a query param like `assigned` or
`milestone` and you were placed in an OAuth redirect, your query params would
be forgetten after returning from OAuth redirects.

This stores the original URL in the session so that when the user gets back
from an OAuth redirect, the app redirects them to their original URL and not
just '/'.