Fix deprecated express middleware uses
This adds more explicit values to `express-session` and removes the use of the
deprecated `req.param` method.

`req.param` is deprecated because it searches for a value from the
* route matching
* query params
* request body

There are individual functions to use for each of those, so it's recommended to
use those instead of the larger generic one. This was used for something that's
always a query param, so the usage is switched to `req.query`.