Issues: fill in repo column
Add the "repo" property to the Issue model and fill it in.

We have three different data sources (API, webhooks, DB). API
responses don't have "repo", webhooks have "repository.full_name" and
the DB has "repo". Here we add code to ensure the data sources all end
up in the model as "repo".

Add code to inject a "repository" attribute for API responses
that don't have them. We have to do it here, cause later (in
"parseIssue") we don't have a reference to the repo.