Pull model: move API construction to Factory
Instead of forcing all instantiation to use the format of the API,
let's just make several factories. Both factories should produce
the same internal structure, but from different data sources.

Previously DB data was transformed into API data and then transformed
into the internal stae via the constructor. Let's just shorten that
to two factory methods that both compose the internal state.

It'd be better if the internal state were a bit more rigid, maybe
in the future. This change shouldn't affect functionality.

Note: Previously, dates were created in the pull object through these
two layers. This was unintentional, but worked because a toString()
on the date was a valid input to fromDateString().

    utils.fromDateString(utils.fromUnixtime(pull.created_on))

Now they are done separately.