Add repo field to git manager and refresh
Updates all calls to the git-manager interface that sits in front of the
github api. Adds the ability to specify a repo to every method that
accepts one which is passed in to params to create the argument object
passed into the github api library.

Adds the forEachRepo method that does a map on whatever function you
pass into it for every repository. The method makes the assumption that
the first parameter of the function will always be the repo parameter.

The forEachRepo method will be used where we want to get every type of
one thing, but now for multiple repositories.

The idea for forEachRepo is to always return an array of things, but now
the array contains items from all the repositories that pulldasher is
watching.