Update primary key to add repo
Since adding multiple repos allows a pull or issue to be duplicate
even if they have the same number, we need to identify them by
both their repo and number. Updates the schema to match the migrations
that were added to drop and add the new primary key spanning repo,
and whatever previous primary key.