Update models to use repo identifier
All the queries that grab stuff from the database need to use repo
for multiple repositories.

Moves the param order of (repo, ...otherthings) so that forEachRepo
can be used in dbManager. This also just makes it easier to read
if everything always has a similar param order.

Moves forEachRepo to utils so that it can be used in both dbManager
and gitManager.