Issues: s/full_name/repo/
Problem: the webhook response has repo information *outside* the
issue body and the getIssue API doesn't have repo information at
all. Our attempt at normalizing the api response needs to modify the
repo data from the webhook (bring it inside the issue)
AND the API (add it).

Previously, the issue object passed from the webhook to
Issue.getFromGH(body.issue) didn't have .repository cause that was on
body.repository, not body.issue.repository.

We need the repo information and it wasn't present where we needed it
in either the webhook nor the api response. So here we add it at
issue.repo so it's in the same spot for both data sources.