Allow null values in repo_name so code can be removed
Code that references repo_name can be removed after this migration is
ran so that when the ORM tries to insert, mysql will fill all the values
with null since they won't be provided by the code.

Since repo_name is no longer going to be used, repo should have the
index.