Update function calls to add in repo argument
Many places where the function definition was updated but the actual
arguments in callsites were not added needed to be fixed up. Throws in
the repo argument in front of whatever was already there.

The dirtyPulls array was indexed based on pull number, but since this is
now multiple repositories, just the number is not unique. Indexes
dirtyPulls by repo plus an octothorpe to separate the number.