Update the pulls variable with pulls from the db
The pulls were not having their signatures updated in the UI because the
pull objects passed to the frontend were out of date.

This was grabbing the pull from its own storage, which then it was
updating with, since this pull is old and stale, it didn't actually
update anything.

Correctly use the dbManager to grab the pull from the database which
will have all the updated data and then it will actually update the
pulls correctly.