Change CREATE TABLE to CREATE TABLE IF NOT EXISTS
This is really helpful if you're building the project using
docker-compose, which requires the database schema to be set up
idemptotently. This way if it's the first or nth time the project is
brought up with docker-compose, the entrypoint script will create the
necessary tables if they're missing, and will ignore them if they
already exist.