Use the full repo name now that repo_name is gone
repo_name was replaced with repo, so these links were breaking. Using
the full repo name can take place of owner + / + repo name since it
already has a slash.

This also grabs the repo from the base branch and not the head.

The frozen link is only going to go to the first frozen pull's repo,
which is not ideal, it will have to be fixed later.