Add context column migration and schema
The context column will represent the github pull status context which
is all just aname for a build status.

In iFixit's case, we use a suite of tests that are separated into main
categories such as Unit, Integration, Image Diff and so on. An example
context for a build might be Unit or Integration.