Update config to use an object with the repo and required builds
The required passing builds addition is a list of all of the builds that
need to have been both run on the pull and have a passing status. If
there is a build that is failing but isn't required then the commit
should should be marked as passing.

The original string that was the repo is now in the full_name property
since this is what github names it.