Work with changed repos config structure
repo_name now has to be grabbed out of the repo object for each repo
instead of just using the repo as a string. Adds a utils function since
it was a prety common task to get out all the repo names, and then use
them in the existing higher order functions.