Add passing function to pull on whether it passes CI
A pull is passing if it can get a successful build for every single
required build in its config. This takes all the required builds and
returns true if there are no builds thare are required that aren't
passing, and false if there are builds that are required that are either
not there or not passing.