Add webpack config
This webpack config builds all javascript files used on the frontend
into a bundle called bundle.js. raw-loader is used so that we can load
the text value of some html files into strings for templating int
javascript.

The ProvidePlugin is necessary because bootstrap assumes that there'll
be a global jquery object.