Remove references and schema of repo_owner
The base branch owner is implied to be the owner of the head branch,
which isn't a correct assumption, but this doesn't change that
assumption, it will have to be fixed another day.

Removes repo_owner from the schema since it was already dropped with
a migration in a previous commit, then stops inserting it with the
db_pull model.