Retrieve the repo of the pull from the base, not the head
The pull's repo is the base commit, not the head, since the head can be
in different repos. the base is always the branch the pull request is
for.

This fixes the problem where pulls were having their repo set as
whatever was the head branch, which is a problem when you try to make a
cross repo pull.