Don't automatically refresh open pulls on start
This was causing the pulldasher app to go over its rate limits pretty
frequently. It was extremely problematic since it would crash on hitting
the rate limit, and immediately crash into the rate limit again.

When there are a very large number of repositories and thus many items
pulldasher needs to lookup, it approaches the rate limit very quickly.