ONly do a single negation for passing builds check
A double negation meant that if there was a build that was required but
was not passing, it would be true since there would be a single build
not accounted for. The single negation is correct, because only when
there are no more builds that need to be accounted for in the array,
then will it be true. since !(empty array) = true = passing.