Use bluebird promises
Bluebird is a pretty performant promise library that implements support
for things like .done and Promise.promisfy that we currently use in
Pulldasher. It's written in a way that's supposed to use less memory
than standard ES6 promises, so I thought it'd be interesting to try it
out.

From some light testing, running pulldasher with then/promise runs at
about 35MB, but running pulldasher with bluebird runs at about 23-25MB.