Update octokit underneath git-manager
git-manager exports its own API to pulldasher, so this updates the
underlying github SDK underneath it, without changing the interface.

Once this is stable for some time we can revisit whether or not to keep
our own second layer of API around octokit, or just use octokit directly
in pulldasher.