Remove jshint and add eslint
We're starting to use more future-y JS syntax and features, and jshint
doesn't seem to support or recognize those. eslint does, so it doesn't
hurt to move over to that and remove one more thing from our Grunt
config.

This also uses lint-staged to check git staged files on commit and husky
to set up the git commit hooks.