Convert theme colors to use CSS variables
The current theme system defines a range of colors and less variables
and builds an output file for each theme. This makes it hard to use with
webpack, because ideally all of the themes are loaded and we can switch
between them using JS instead of pulling in a new theme.

By using native CSS variables, we can define a set of color values
encapsulated by theme name, and dynamically switch which variables are
in scope. We're also no longer duplicating styles and just shipping a
list of variables instead of two build outputs.

All style is imported by a new main.less file which is specified as an
entrypoint in our webpack config, and the style is added to the DOM on
page load thanks to style-loader.