Template rendering: rename debug mode
Passing through "debug" to the .render() function triggers
rendering the view in debug mode, so let's not do that.

Debug mode apparently spints out the compiled template function
for the given template to stdout (on every call). Spitting out
a JS function on stdout on every request was useless, drop it.