fix(authentication): deduplicate req.logIn function
This also adds a catch handler for the org membership check promise and
passes the error to `next` to be handled. Currently this goes to the
default error handler, but before this the unhandled promise rejection
would trigger a warning but the response wouldn't end because the
middleware chain wouldn't continue.

Now the error message is printed out and the response is ended
successfully.