Github API: get combined view of github statues
The previous endpoint listed every status ever on that commit
in chronological order. We're only interested in the most recent status
for each context. Otherwise, we'd have to sort them and collapse them
per context to get the most recent state for each context.

Github's api does this for us.

Also, add a few logs.