config.repos: normalize repo list on load
In a previous pull, config.repos learned to handle an array of objects
instead of just an array of strings. Not all usages of the config.repos
setting were updated to handle the different ways of representing the
data.

Instead of handling both formats in each usage, let's normalize when
loading the config.