Github Pull hooks: refresh on more events
For adding or removing a milestone to a pull, GH sends an "issue event"
cause every pull is *also* an issue.

Previously, we tried to use the body to update the issue in the DB
(which didn't exist). Now any time we get an issue event for a pull,
we just refresh that whole pull. We have to do this cause the body that
comes with issue events is incomplete.