Frontend: Connect to backend via socket.io
Only attempt connecting to the backend if the list of dummyPulls is
empty. When in dummy-mode, we just use the dummy pulls from the json
file, allowing development in isolation from the backend and its
dependencies.

Re-use the socket token auth mechanism that associates an anonymous
websocket connection with authenticated page load with a passed-through
token. This may be replaced with getting the auth cookie from the
websocket connection directly in the future.