Pulls: Add age indicator
We want to be able to tell how hold a pull is at a glance. This adds the
indicator as well as adds it to the pull element.

QA:
--
I currently have it set up on https://pulldasher-dev.cominor.com.  Play
around with it! You could open a test pull to ensure it adds the proper
dots.