Feature: Add GitHub Action support
While switching over to GitHub Actions on iFixit, we realized that
the commit statuses were not showing up on the pull cards. After some
digging, we realized that GitHub Actions are actually `checks` which we
need to pull in using a different endpoint.

We want pulldasher to work with commit statuses as well as Actions. To
do this, we just concatenate the statuses and the checks. When we create
the Status in git-manager, we just pull off the relevant information.
and map the check values to status values. This let's us keep most of
the other code the same as before.