pulls-state: move init protection into use hook
Let's leave all the double-initilization protection jut in the caller
instead of splitting it between the caller and callee.

Also, add back the `useEffect()` because this function will be executed
more than once (each time react re-renders the app. We just want to
protect against using it in more than one component.