package-lock: install pacakges dropped in merge commit
There was a merge conflict in the package-lock (previous commit).
I resolved it by just accepting masters package lock. I then re-did the
changes (npm install list-of-packages) and this is the result.

I also dropped a prop that typescript was complaining about, cause Flex
doesn't take a `size` prop. The history doesn't say why that was in
there anyway.