Frontend: add configurable sorting
The default sort is sensible, surfacing pulls that need action
over those that don't.

I opted to abandon the score-based sorting for more explicitly
ordered lists. There will be less mystery and we have few enough
pulls that it really doesn't matter.

Also, I'm not sure if we'll want configurable sorting, we might
want to drop it, but I modeled it after the filtering.