Cryo Pulls: fix toggle!
There was an off-by-bool bug where the now-out-of-date showCryo value
was used to update the filter, leading to it workin in the opposite way.

Let's pull the setPullFilter() call out and have it more
deterministically set from the value of showCryo.

Also, add `useCallback()` to prevent a new event listener each time the
navbar re-renders.