frontend: don't mutate pull.show
mutating pull.show requiring consumers to extract pull.show and send it
along as a separate prop if their component is memoized.

Because of this gotcha and others, it's generally not a good idea to mutate state
directly. It's better to treat the top-level state objects as immutable.

Instead of that pattern, let's try something new: providing the Set of
pulls that pass filters as a separate value in the pulls context.

This gives the same speed benefits (cause pull-cards are still just
hiding themselves) but we aren't mutating state in the pull object,
requiring consumers .