Frontend: make refresh *pretend* when using dummy pulls
If we're running without a real backend (i.e. using an array of dummy
pulls) then we need to fake the backend (at least in the case of th
'refresh' command.

Specifically, we need to wait a bit, then pretend like a new version of
the pull came from the backend.