Backend: load all open pulls from the DB
Previously, we only loaded pulls that were in our *repo* array.

This meant that after a restart, some pulls may disappear.
We may receive webhooks for repos not in that array,
show the pulls for a while (cause they are in memory), then they
disappear on restart cause they don't get loaded from the DB.

The repo list is now only used on start when we refresh all pulls from
the API (need to know which repos to ask about).

=== Future Work ===
We may want to support '{orgname}/*' in the repo-list or something.