Frontend: Hide Closed Pulls
The backend won't send us closed pulls in the initial list, but we
hadn't been hiding pulls that *become* closed after the page has been
loaded. This meant that a merged pull would jut sit around in the ready
column, even after it's been closed.

The old UI did this, it was just overlooked.