use-url-state: avoid empty render
Previously, this would render once with `state == null`, then after
mounting, it'd run the useEffect() and parse the state out of the url.

The url is available immediately, so let's parse it out in the "get the
default" function and avoid the double render. This is important
functionally for later components that store more complex state in the
url.