use-url-state: fix bug when going back to first page
The internal "state" objects sent along with each `pushState()` help us
add more complex state to history navigation. Each pushState() call
takes a new url and a state object. These state objects are handed back
to you on popstate events (the back button).

Pages (regardless of the url) load with an empty state object. So, when
a user navigates *back* to the initial page, the state object is empty,
even if the url isn't. This meant that our use-url-state hook was
setting the current value to the default when you went back to the
inital page, even if there was important state in the url.

Here we change to use parsing the url as a fallback if there's no info
in the state object.