filter-menu: unbreak history navigation!
We only set the filter based on the "selected values" when a menu item was
clicked or on the first render. This of course excludes history
navigation when selectedValues may change cause the user went *back*.

Changing this to a declarative style with useEffect() (filter is determined by
selectedValues) makes the code simple and removes the bug.

Also, the problem with typescript thinking onChange might be called with
single strings instead of an array seems to have disappeared, so we
don't need the onSelectedChange() function.