Reviews: generate links in frontend correctly
Previoulsy, we assumed all signatures were made from issue comments.
Now we generate the right link depending on the sourec_ttype of the
comments.

For now, let's just compare ids to determine if the signature is from a
review. Their numbers are high enough that the chances of overlap (same
pull