Required Statuses: include "missing" CI bubble
If a pull doesn't have a required status, include a grey CI bubble
like the other CI statuses that says "Missing" to make it obvious which
one can't be found.

Think of it as a placeholder for an unmet requirement. Without this,
it's not obvious which of ~13 required statuses isn't present.