CR Leader List: have it show unique CRs in last 14 days
The Leader list was not that helpful. It was showing the total
CR signature count across all pulls in pulldasher's memory (not the DB).
This wasn't that useful of a metric because that was an abritrary number
of pulls and 7 CRs on one pull isn't the same as 1 CR on 7 pulls.

Now:

* Only include open pulls and those merged less than 14 days ago.
* Only include signatures made int he last 14 days
* Count only the number of pulls each user has CRed (not the number of
  signatures)

Future work:

* Change pulldasher to load even closed pulls from the DB so
  these numbers stay the same on restart.
   * Right now it only loads open pulls on restart
* Maybe: drop list and only show how you compare to mean or median of the
  group.