Author/Repo Filter: indicate filtering
Previously, there was no indication that a filter was applied until you
clicked the button. It was easy to miss that there was a filter and that
you weren't seeing all the pulls.

Now, we change the button to be filled (instead of just an outline) when
a filter is present.