Commit Statuses UI: show popup for more deets
We have a *lot* of commit statuses in our main repo now, (16 or so).
This is far too many to effectively display and make clickable on the
left side of each pull.

Solution
========
Show one status bubble for each state represented, scaled roughly by how
many have that state. Make that whole section clickable and render a popup
with more details.