filters: use function for of setState() to fix bug
Previously, we used the `filters` from the closure in our returned
setFilter(). This meant that consumers had to use a new "setter"
function each time `filters` changed. This is fairly unusual for a
useState()-like hook, most have a stable `setState()` function.

Because of this, there was actually a bug down the line. filter-menu
assumed the setFilter() function didn't change and effectively memoized
it. Un-memoizing caused infinite re-renders and the easier fix was to
use the functional form of `setState()` that provides the old value
and let's you mutate it.