Commit Statuses: fix scaling issues
flexBasis is definitely not what we should use here, it sets the
pixel-size of elements (or at least their "desired size").

flexGrow is effectively a weighting telling the browser how to
distribute extra space. This accomplishes our job of scaling
the bubbles based on how many there are in each status much better.