Signatures: Only show one bubble per requirement
In #298 we decided to only show one CR / QA bubble per requirement but
to include more detail and links in a popup.

Here, we do that, sorting sigs so that the user's own sigs show up
first and then only showing one bubble per required sig, even if there
are many active and inactive sigs.

Effectively, if the current user has an active sig it'll be
shown, and they have an inactive sig, it'll be shown before any other
inactive sig.

a