notifications: Only ignore PRs we can see
This ensures that we don't leave a PR ignored once it drops out of the
class for which the user should be notified. In particular, we want to
re-notify the user when a notification was addressed (causing the PR
to leave the queue) and then a change was made to the PR, causing it
to be ready for review once again.