Notifications: Only track relevant pulls in each column
Previously, we were including all of the user's PRs in the list which
would trigger notifications in the CR or QA column. We don't really
want that, for two reasons:
1. It notifies when they go from CR-complete to needs-CR
2. It doesn't notify when a self-QA expires and needs to be updated.

This now splits out two hooks: one for pulls which need re-review, and
one for pulls which are mine. We then use those as appropriate to
trigger the notifications.