remove buttons, use menu with checkboxes
This commit removes the two buttons in the navigation bar and replaces
them with one menu ("Label").

This menu contains a checkbox group of two checkboxes (one for each
special label - external_block and Cryogenic Storage). These checkboxes
are held in stack and have similar styling properties to the
MenuItem and MenuItemOption.