Reviews: Automatically get CR status from Review status
Reviews are awesome in GH. They have natively defined state in GH which
we can read in pulldasher. Let's pull that information automatically
from the Review.

The old `CR :emoji` should still work on regular comments. We plan to
phase this out if we like working strictly with Reviews for the CR
status.

Note: We now don't listen to any other signatures on the Review comment.

QA:
--
Check out this branch on pulldasher-dev and play around with it.