Notifications: Add click action to open pull
This makes it much more convenient to open the pull the notification
is about.

As part of this, I changed the notification to only show for the first
pull when there are multiple pulls to notify about. This seems to me
to make sense, because the notification is generally truncated enough
for me that it doesn't help to have the full list, and I think it
makes more sense to do this than to show the list of pulls and then
only link to one of them. Also, the vast majority of the time, the
notification is for a single pull, so this won't actually change
behavior in any interesting way.