Notifications: Use redirect to allow more control
This tries opening a second Pulldasher window which then redirects to
a specified URL, in order to allow our JS to try to give that window
focus.

This idea was originally intended to allow us to use the return value
of `window.open`, but since that didn't work, I tried handling the
focus in the redirect JS. That also doesn't seem to really work.